Moved one-on-one chat UI classes to a UI folder
