+++ app/src/main/java/com/courseproject/tindar/ui/matchlist/MatchListFragment.java
@@ -13,7 +13,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.courseproject.tindar.BlankNavViewModel;
-import com.courseproject.tindar.ChatActivity;
+import com.courseproject.tindar.ui.chat.ChatActivity;
 import com.courseproject.tindar.R;
 import com.courseproject.tindar.controllers.likelist.LikeListController;
 import com.courseproject.tindar.ds.DatabaseHelper;
